Align mop for new Montage and Mr boot scripts
:warning: This introduces backward-incompatible changes that should be
coordinated with a major release of `mr` and `montage`.

Montage and Mr are undergowing an overhaul of their bootstrapping
process in their respective `build` branches.  This overhaul entails
using a build script to generate their `montage.js` and `boot.js`
scripts.  These are the corresponding changes needed to support bundling
the new bootstrapping processes.