Fix shared packages bug
Fixes a bug that was preventing a correct application optimisation under
certain circumstances. When a project packages were using the same
packages that Montage does, the order in which Mop would scan and load
files would affect the output optimisation, in some cases resulting in
an invalid one that would prevent the application to work.