Fix missing css bug and add warning
Fixes a bug happening when trying to embed a file from a link rel
stylesheet tag. If the href path had node_modules on it, the file could
not be found because of the bug, it was not embedded and the link tag
ended being removed.

Unrelated to the bug or bug fix, this commit also adds a warning for
missing stylesheet files that are not found/existent.